Edited index, attractions and bookings page
Index page - Made the quick info columns line up to the carousel borders

Attractions page - Added ride requirements, legends to the park map

Bookings page - Added margins to the tickets so that it does not look too close to the section above and the footer. Also Improves on the mobile view.